+++ example/lib/livePlaybackWithoutBackPressure/live_playback_without_back_pressure.dart
@@ -153,7 +153,7 @@ class _LivePlaybackWithoutBackPressureState
       //bufferSize: cstBLOCKSIZE,
     );
     feedHim(data);
-    _mPlayer?.logger.d('Finished');
+    _mPlayer.logger.d('Finished');
 
     setState(() {});
   }