+++ example/lib/recorder_onProgress/recorder_on_progress.dart
@@ -18,7 +18,6 @@
  */
 
 import 'dart:async';
-import 'package:audio_session/audio_session.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_sound/flutter_sound.dart';
 import 'dart:typed_data';