+++ example/lib/streams/streams.dart
@@ -19,7 +19,6 @@
 
 import 'dart:async';
 import 'dart:typed_data';
-import 'package:audio_session/audio_session.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_sound/flutter_sound.dart';